pyproject.toml
[tool.poetry]
name = "golld"
version = "0.1.0"
description = "GoLLD: Glide on LLD"
authors = ["Diwank Singh Tomer <diwank.singh@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.12.0"
Pillow = "^9.2.0"
accelerate = "^0.10.0"
datasets = {extras = ["vision"], version = "^2.3.2"}
diffusers = "^0.0.4"
torchvision = "^0.13.0"
tqdm = "^4.64.0"
transformers = "^4.20.1"
pandas = "^1.4.3"
huggingface-hub = "^0.8.1"
bitsandbytes-cuda102 = "^0.26.0"
einops = "^0.4.1"
bitsandbytes-cuda113 = "^0.26.0"

[tool.poetry.dev-dependencies]
ipython = "^8.4.0"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
